requirements.txt
# Python 3.9+ (recommended on Apple Silicon — no Python 3.6 builds for osx-arm64)
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.62.0
ruamel.yaml>=0.17.0
torch>=2.0.0
torchvision>=0.15.0
# Optional: wandb
